Problem: #69 App architecture inefficiencies and overhead
Solution: tried clearing server folder. #69
