+++ imports/api/activityLog/server/methods.js
@@ -1,5 +1,6 @@
 import { Meteor } from 'meteor/meteor'
 import { ActivityLog } from '/imports/api/indexDB.js'
+import { log } from '/server/main.js'
 
 Meteor.methods({
   sendMessage: function(userId, message, from) {
@@ -22,4 +23,27 @@ Meteor.methods({
       })
       }
     }
+});
+
+Meteor.methods({
+    'markNotificationsAsRead': function() {
+        if (!Meteor.userId()) { throw new Meteor.Error('error', 'please log in') };
+
+        ActivityLog.update({
+            owner: Meteor.userId(),
+        }, {
+            $set: {
+                read: true
+            }
+        }, {
+            multi: true
+        }, function(error) {
+            if (error) {
+                log.error('Error in markNotificationsAsRead', error)
+                throw new Meteor.Error(500, error.message);
+            }
+        });
+
+    }
+
 });
