+++ imports/startup/server/index.js
@@ -1,6 +1,7 @@
 import '../../api/bounties/server/methods.js'
 import '../both/index.js'
 
+//publications
 import '/imports/api/bounties/server/publications'
 import '/imports/api/hashing/server/publications.js'
 import '/imports/api/miscellaneous/server/publications'
@@ -14,12 +15,16 @@ import '/imports/api/coins/server/publications'
 import '/imports/api/redflags/server/publications'
 import '/imports/api/activityLog/server/publications'
 
+//server methods
 import '/imports/api/coins/server/methods'
 import '/imports/api/wallet/server/methods'
 import '/imports/api/activityLog/server/methods'
+import '/imports/api/developers/server/methods'
+import '/imports/api/users/server/methods'
 
 // import './fastRender.js'
 import './fixtures.js'
 
+// collection configs
 import '/imports/api/users/users.js'
 import '/imports/api/coins/currencies.js'
