--- server/connections.js
@@ -1,7 +1,0 @@
-  Meteor.methods({
-      getUserConnectionInfo: function() {
-          
-              return this.connection;
-          
-      },
-  });
